requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
cloudinary>=1.40,<2.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<24.0
